dev-python/tzdata: use setuptools to build instead of hatchling
hatchling is meme software and keeps breaking.

The thing we are trying to do is explicitly supported in the setuptools
documentation, which is a strong commitment that they intend to support
this use case:

```
If you want to publish a distribution that does not include any Python
module (e.g. a "meta-distribution" that just aggregate dependencies),
please consider something like the following:

[tool.setuptools]
packages = []
```

Additionally, hatchling depends on a ton of packages, including...
setuptools. This is an additional 6 packages to bootstrap, and
setuptools is significantly more universal anyway.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>