sys-libs/musl: drop the need for USE=split-usr
We do this similar to how glibc works:

- /lib64/ld-linux-x86-64.so.2 is the canonical loader

- /lib64/libc.so.6 is the canonical C library

- /usr/lib64/libc.so is needed by the compiler, and is a linker script
  which loads the canonical loader and C library from /lib64, which may
  or may not be symlinked to /usr

For musl, this is a bit hairier. Both libc.so and ld.so are the same
file, and ld.so is a symlink to libc, which seems backwards. In
comparison to glibc, no versioned name for libc exists at all.

Changes:

- Reverse the order around so that ld.so is the real file.

- No canonical C library is needed, since it is satisfied by ld.so; we
  don't need libc.so for this purpose.

- libc.so is needed by the compiler; change it from a real file, to a
  linker script which loads the canonical loader-slash-C-library.

On merged-usr systems, both the loader and the C-library-linker-script
are in /usr/lib. On non-merged-usr systems, the loader is in /lib and
the C-library-linker-script is in /usr/lib next to libc.a. No conflicts
occur between /lib and /usr/lib, and the contents of the package are now
identical regardless of the value of USE=split-usr.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>