WIP: app-admin/sudo: manually hook up CFLAGS to LDFLAGS
Usually build systems do this themselves. It is necessary to pass at
least some compile flags to the link stage, for LTO, since lto-wrapper
reprocesses the object files and wants the original CFLAGS in the
process.

This is detectable because -frecord-gcc-switches disappears and QA
warnings result. (And sudo doesn't pass CFLAGS to its manual `libtool`
commands hardcoded in Makefile.in)

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>