dev-python/scipy: fix src_configure to handle filter-lto correctly
Use of python_configure_all is a bit broken, because distutils-r1 is a
bit broken. It has the intriguing value-claim that *FLAGS shall be
modified with local -x as part of run-phase, which means that all
modifications are reset as soon as any given phase exits. Including
sub-phases. If you make changes in python_configure or
python_configure_all, as you are expected to do instead of defining
src_configure and then running the eclass phases manually, your changes
inherently get erased before you actually *get* to running builds (i.e.
reading the variables).

Capitulate to distutils-r1 variable scoping purity demands, by declining
to use distutils-r1 for this kind of work. Everyone wins. :)

Note: it "worked" anyways prior to reworking meson.eclass to pass
-Db_lto based on `is-flagq flto`. This is because filter-lto removed it
from LDFLAGS, and since distutils-r1 doesn't localize that variable, the
changes stuck around. As a result:

- the previous fix to scipy caused scipy to be compiled with LTO, but
  not linked with LTO
- meson.eclass changes caused scipy to be built with -Db_lto=true, which
  affects both compiling and linking
- this patch causes scipy to be built with -Db_lto=false

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>