sci-libs/hdf: drop painful dead weight
I am not fighting with Modern C more than I have to, so just prune this.

The key here is that upstream hdf4 doesn't support it and doesn't want
you using it, because it appears to just be an uplifted copy of netcdf
code. So why not, erm, just use netcdf?

Good question, and this package doesn't have an answer to that. Yeet
these vendored tools into the sunset. No need to rename them to private
copies.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>