sci-geosciences/opencpn: remove mysterious double-eapply
In multiple phase functions, "default" is run in addition to cmake_*
eclass phases. This is entirely redundant, as cmake.eclass does this for
you. As a result, running it twice would simply attempt to apply a patch
that has already been applied, inevitably failing, and causing a very
confusing build error when you know the patch does indeed apply.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>