net-libs/gtk-vnc: bind tightly to whether gtk was built with wayland
The upstream code checks whether GDK_WINDOWING_WAYLAND is defined by the
gtk headers, and if so will compile against the wayland symbols it
provides. This means that gtk-vnc built on a system with gtk+[-wayland]
will be compatible with anything, but when built on a system with
gtk+[wayland], requires that at runtime.

This cannot be expressed with USE flags, so we do the next best thing
and arbitrarily restrict gtk-vnc[-wayland] to only build and install on
a system with gtk+[-wayland].

Bug: https://bugs.gentoo.org/624960
Closes: https://bugs.gentoo.org/865659
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>