x11-libs/wxGTK: bind tightly to whether gtk was built with wayland
The upstream code checks whether GDK_WINDOWING_WAYLAND is defined by the
gtk headers, and if so will compile against the wayland symbols it
provides. This despite that wxGTK has a wayland USE flag, and checks it
in the build system -- just not for this.

This means that when built on a system with gtk+[-wayland]
will be compatible with anything, but when built on a system with
gtk+[wayland], requires that at runtime regardless of USE flag.

This cannot be expressed with USE flags, so we do the next best thing
and arbitrarily restrict wxGTK[-wayland] to only build and install on
a system with gtk+[-wayland].

Bug: https://bugs.gentoo.org/624960
Closes: https://bugs.gentoo.org/927952
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>