xfce-base/libxfce4ui: bind tightly to whether gtk was built with wayland
The upstream code checks whether GDK_WINDOWING_WAYLAND is defined by the
gtk headers, and if so will compile against the wayland symbols it
provides.

This means that libxfce4ui built on a system with gtk+[-wayland]
will be compatible with anything, but when built on a system with
gtk+[wayland], requires that at runtime regardless of USE flag.

This cannot be expressed with USE flags, so we do the next best thing
and arbitrarily restrict libxfce4ui[-wayland] to only build and install on
a system with gtk+[-wayland].

Fixed upstream in:
https://gitlab.xfce.org/xfce/libxfce4ui/-/commit/967359624e408365b4c027bab66e737063af604b

Bug: https://bugs.gentoo.org/624960
Closes: https://bugs.gentoo.org/873520

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>