dev-python/tlsh: actually wire up testsuite
Upstream has a testsuite, it's just in a different directory compared to
the python module itself. It internally uses the script we used to
crudely run, but it also actually checks the results against fixtures.

As a result, it checks more than just "smoketest: does the module import
and execute its functions without literally crashing".

It also doesn't redirect the testsuite stdout/stderr to devnull, since
it now actually does something constructive and we'd like to see the
status from it.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>