gnome-extra/cinnamon-control-center: avoid automagic deps on gtk+[X,wayland]
Currently, one is forced to compile c-c-c and gtk with the same value of
USE=wayland which makes it difficult to provide generically reusable
binaries. It is better than automagically depending on wayland and then
lacking it, but not ideal. Preferable would be for c-c-c[wayland] to
require gtk+[wayland] but when built without wayland support, to not
care whether gtk itself is.

We now support a mechanism for packages to opt out of gtk providing the
API, for use by packages that don't provide a `-D wayland=true` but
autodetect whether gtk was built with it. Use that mechanism.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>