python-utils-r1.eclass: Depend on dev-lang/pypy directly
Depend on `>=dev-lang/pypy-3.10:=` rather than the backwards
compatibility `dev-python/pypy3` package.  Note that the package needs
to remain at least for some time after the next subslot bump, so that
users rebuild all packages and get the updated dependency across
the system.

Note that this requires pkgcheck to be updated first.

Signed-off-by: Michał Górny <mgorny@gentoo.org>